requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
boto3==1.24.13
botocore==1.27.13
braintree==4.15.2
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.2
defusedxml==0.7.1
Django==4.0.5
django-ckeditor==6.4.2
django-cors-headers==3.13.0
django-environ==0.9.0
django-js-asset==2.0.0
django-storages==1.12.3
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.4
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
Markdown==3.3.7
MarkupSafe==2.1.1
mutagen==1.45.1
oauthlib==3.2.0
Pillow==9.1.1
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.1.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.1
requests==2.28.0
requests-oauthlib==1.3.1
s3transfer==0.6.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
stripe==3.4.0
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
whitenoise==6.2.0
zipp==3.8.0
